feat: consolidate workers into v2/backend with Sleeper data UI
- Move Temporal workers (activities, workflows, schedules, sleeper client)
  from root workers/ module into v2/backend so Go models are shared
- Add cmd/worker and cmd/seed binaries to the backend module
- Add Sleeper GORM models to AutoMigrate in database initialization
- Add GET /api/sleeper/stats, /trades, /drafts endpoints
- Add Sleeper stat cards on homepage (leagues/trades/drafts counts)
- Add /sleeper/trades and /sleeper/drafts paginated list pages
- Add Trade Data and Draft Data links to the nav header
- Update Dockerfile to build both server and worker binaries,
  running the worker in the background via an entrypoint script
- Delete root workers/ directory
