--- workers/internal/db/migrate.go
@@ -1,19 +1,0 @@
-package db
-
-import (
-	"database/sql"
-	"embed"
-
-	"github.com/pressly/goose/v3"
-)
-
-//go:embed migrations/*.sql
-var migrationsFS embed.FS
-
-func RunMigrations(sqlDB *sql.DB) error {
-	goose.SetBaseFS(migrationsFS)
-	if err := goose.SetDialect("postgres"); err != nil {
-		return err
-	}
-	return goose.Up(sqlDB, "migrations")
-}
