--- workers/internal/db/postgres.go
@@ -1,22 +1,0 @@
-package db
-
-import (
-	"fmt"
-	"os"
-
-	"github.com/joho/godotenv"
-	"gorm.io/driver/postgres"
-	"gorm.io/gorm"
-	"gorm.io/gorm/logger"
-)
-
-func Connect() (*gorm.DB, error) {
-	_ = godotenv.Load()
-	dsn := os.Getenv("DATABASE_URL")
-	if dsn == "" {
-		return nil, fmt.Errorf("DATABASE_URL not set")
-	}
-	return gorm.Open(postgres.Open(dsn), &gorm.Config{
-		Logger: logger.Default.LogMode(logger.Silent),
-	})
-}
